+++ NEWS.md
@@ -30,6 +30,9 @@ variadic inputs.
 have been removed. The functionality of these can be achieved with the existing
 members.
 
+- `r_sexp::length` has fully deprecated and has been removed. 
+Use `cppally::length` for returning object length.
+
 ## r_function
 
 - New class `r_function` to safely call R functions from C++.